Move setup implementation around.
